third_party/blink/web_tests/fast/events/window-onerror-sandbox-01-expected.txt

CONSOLE ERROR: Uncaught ReferenceError: hahaha_good_luck_finding_me is not defined
This test should trigger 'window.onerror' inside the sandboxed iframe, and successfully handle the error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS message is "Uncaught ReferenceError: hahaha_good_luck_finding_me is not defined"
PASS successfullyParsed is true

TEST COMPLETE
